read notif arch
